requirements.txt
openai>=1.0.0
pyyaml>=6.0
tqdm>=4.64.0
pandas>=1.3.0
openpyxl>=3.0.0
tenacity>=8.2.0 